My own take
removed the bib, since it was not working
now uses tor.
changed user-agents for every call
uses tor to avoid captcha
